Rework and deprecate ProfileMainWindow
